saml-inspector/requirements.txt
Flask==3.0.3
pysaml2==7.5.0
gunicorn==23.0.0
httpx==0.27.2
